code changes to run rosR on 32bit systems

+++ src/PublisheR.cpp
@@ -23,8 +23,8 @@ void rros_stream_write_int16	(OStreamR *s, signed short	val) { s->next(val); }
 void rros_stream_write_uint16	(OStreamR *s, unsigned short	val) { s->next(val); }
 void rros_stream_write_int32	(OStreamR *s, signed int	val) { s->next(val); }
 void rros_stream_write_uint32	(OStreamR *s, unsigned int	val) { s->next(val); }
-void rros_stream_write_int64	(OStreamR *s, signed long	val) { s->next(val); }
-void rros_stream_write_uint64	(OStreamR *s, unsigned long	val) { s->next(val); }
+void rros_stream_write_int64	(OStreamR *s, signed long long	val) { s->next(val); }
+void rros_stream_write_uint64	(OStreamR *s, unsigned long long	val) { s->next(val); }
 void rros_stream_write_float32	(OStreamR *s, float		val) { s->next(val); }
 void rros_stream_write_float64	(OStreamR *s, double		val) { s->next(val); }
 void rros_stream_write_string	(OStreamR *s, char*		val) {
@@ -38,8 +38,8 @@ void rros_stream_write_int16_array	(OStreamR *s, std::vector<signed short>*	vec)
 void rros_stream_write_uint16_array	(OStreamR *s, std::vector<unsigned short>*	vec) { s->next(*vec); }
 void rros_stream_write_int32_array	(OStreamR *s, std::vector<signed int>*		vec) { s->next(*vec); }
 void rros_stream_write_uint32_array	(OStreamR *s, std::vector<unsigned int>*	vec) { s->next(*vec); }
-void rros_stream_write_int64_array	(OStreamR *s, std::vector<signed long>*		vec) { s->next(*vec); }
-void rros_stream_write_uint64_array	(OStreamR *s, std::vector<unsigned long>*	vec) { s->next(*vec); }
+void rros_stream_write_int64_array	(OStreamR *s, std::vector<signed long long>*	vec) { s->next(*vec); }
+void rros_stream_write_uint64_array	(OStreamR *s, std::vector<unsigned long long>*	vec) { s->next(*vec); }
 void rros_stream_write_float32_array	(OStreamR *s, std::vector<float>*		vec) { s->next(*vec); }
 void rros_stream_write_float64_array	(OStreamR *s, std::vector<double>*		vec) { s->next(*vec); }
 void rros_stream_write_string_array	(OStreamR *s, std::vector<std::string>*		vec) { s->next(*vec); }
